refactor(#82): fetchApi.ts 추가로 중복되는 선언 제거
