+++ packages/api/src/domains/user/endpoints.ts
@@ -1,3 +1,4 @@
+import { fetchApi } from '../../shared/utils/fetchApi';
 import type {
   LoginRequest,
   LoginResponse,
@@ -6,49 +7,6 @@ import type {
   RegisterResponse,
   UserInfoResponse,
 } from './types';
-
-const BASE_URL = '/api';
-
-async function fetchApi<T>(
-  url: string,
-  options: {
-    method?: 'GET' | 'POST' | 'PUT' | 'DELETE' | 'PATCH';
-    data?: unknown;
-    headers?: Record<string, string>;
-    params?: Record<string, string | number>;
-  } = {},
-): Promise<T> {
-  const { method = 'GET', data, headers = {}, params } = options;
-
-  const fullUrl = new URL(url, BASE_URL);
-  if (params) {
-    for (const [key, value] of Object.entries(params)) {
-      if (value !== undefined && value !== null) {
-        fullUrl.searchParams.append(key, String(value));
-      }
-    }
-  }
-
-  const config: RequestInit = {
-    method,
-    headers: {
-      ...(!(data instanceof FormData) && { 'Content-Type': 'application/json' }),
-      ...headers,
-    },
-  };
-
-  if (data) {
-    config.body = data instanceof FormData ? data : JSON.stringify(data);
-  }
-
-  const response = await fetch(fullUrl.toString(), config);
-
-  if (!response.ok) {
-    throw new Error(`HTTP ${response.status}: ${response.statusText}`);
-  }
-
-  return response.json();
-}
 
 export const userApi = {
   // 회원가입 (POST /users)
